refactor(models): 为多个模型添加注释并优化关联关系
- 为 Alipay、Aliyun、Certificate 等多个模型添加类注释说明
- 优化 Aliyun 模型中域名、DNS、ECS 关联关系的定义
- 为 ApiLog 模型添加 MorphToUser Trait 替代手写 user 关联方法
- 调整 Coupon、Identity、Order、Wechat 等模型的关联关系方法及属性转换
- 修改 OrderItem 的 getSubTotalAttribute 返回类型为 float
- 为多个模型添加属性访问器及修改器注释
- 调整并完善多处方法注释与参数类型声明
- 优化黑名单和敏感词服务，增加方法注释及内部结构重构
- 为多个服务类增加方法注释，完善类型提示
- 移除多余的 PHPDoc 标签及冗余代码
- 统一代码风格，调整空格及缩进，修正部分常量声明位置

+++ app/Models/OrderItem.php
@@ -43,7 +43,7 @@ class OrderItem extends Model
      */
     public function getSubTotalAttribute(): float
     {
-        return bcmul($this->qty, $this->price, 2);
+        return (float) bcmul($this->qty, $this->price, 2);
     }
 
     /**
